+++ Paciente.py
@@ -1,4 +1,5 @@
 import psycopg2
+from psycopg2 import Error
 
 class Paciente:
     
@@ -30,6 +31,10 @@ class Paciente:
 
         """
         Metodo que inserta en la tabla de paciente los datos de un objeto
+        Args:
+            conn: La conexión a la base de datos
+            cursor: El cursor de la base de datos
+        Returns:
         """
 
         query = """
@@ -51,6 +56,11 @@ class Paciente:
     def read(cursor, idPaciente):
         """
         Método selección de la tabla, o lectura de los datos si lo prefieres
+        Args:
+            cursor: El cursor de la conexión a la base de datos
+            idPaciente: El id del paciente a buscar
+        Returns:
+            cursor.fetchone() : La tupla con los datos del paciente
         """
         query='SELECT * FROM paciente WHERE idPaciente = %s ;'
         cursor.execute(query, (idPaciente, ))
@@ -59,6 +69,13 @@ class Paciente:
     def update(self, conn, cursor, paciente):
         """	
         Método que actualiza en la tabla de paciente los datos de un objeto
+
+        Args:
+            conn: La conexión a la base de datos
+            cursor: El cursor de la base de datos
+            paciente: El objeto con los datos actualizados
+        Returns:
+            cursor.fetchone(): La tupla con los datos del paciente
         """
 
 
@@ -89,15 +106,26 @@ class Paciente:
     def delete(conn, cursor, idPaciente):
         """"
         Método que borra en la tabla de paciente los datos de un objeto
-        """
-        print("Datos borrados:")
-        resultado = Paciente.read(cursor, idPaciente)
-        print("Datos del paciente:")
-        print(f"Nombre: {resultado[1]} {resultado[2]} {resultado[3]}")
-        print(f"Teléfono: {resultado[4]}")
-        print(f"Fecha de nacimiento: {resultado[5]}")
 
-        query = "DELETE FROM paciente WHERE idPaciente = %s"
-        cursor.execute(query, (idPaciente,))
-        conn.commit()
-        return cursor.rowcount
+        Args: 
+            conn: La conexión a la base de datos
+            cursor: El cursor de la base de datos
+            idPaciente: El id del paciente a borrar
+        Returns:	
+            bool: True si se borra, False si no
+        """
+        try:
+            if cursor.rowcount > 0:
+                query = "DELETE FROM paciente WHERE idPaciente = %s"
+                cursor.execute(query, (idPaciente,))
+                conn.commit()
+                cursor.rowcount
+                return True
+            else:
+                print("No se encontró el paciente que desea borrar.")
+                return False
+        except psycopg2.Error as e:
+            print(f"Error al borrar el paciente: {e}")
+            return False
+
+
